refactor move installableUtils into utils directory
